// src/interfaces/ILoyalty.tsx
export interface ILoyalty {
  id?: number;
  // eslint-disable-next-line camelcase
  tpg_valuation?: number;
  name?: string;
  // eslint-disable-next-line camelcase
  icon_url?: string;
  // eslint-disable-next-line camelcase
  website_url?: string;
  // eslint-disable-next-line camelcase
  loyalty_unit?: string;
  type: string;
  active?: string;
  // eslint-disable-next-line camelcase
  create_by?: string;
  // eslint-disable-next-line camelcase
  create_date?: string;
  // eslint-disable-next-line camelcase
  update_by?: string;
  // eslint-disable-next-line camelcase
  update_date?: string;
}